requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.2
contourpy==1.2.1
cycler==0.12.1
datasets==2.18.0
dill==0.3.8
docker-pycreds==0.4.0
einops==0.7.0
filelock==3.13.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
gluonts==0.14.4
huggingface-hub==0.22.2
hydra-core==1.3.2
idna==3.7
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
lightning-utilities==0.11.2
lit==18.1.3
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
pandas==2.1.4
pillow==10.3.0
protobuf==4.25.3
pyarrow==15.0.2
pyarrow-hotfix==0.6
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
pytorch-lightning==2.0.4
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
rich==13.7.1
safetensors==0.4.3
scikit-base==0.7.7
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
sentry-sdk==1.45.0
setproctitle==1.3.3
sktime==0.28.0
smmap==5.0.1
sympy==1.12
threadpoolctl==3.4.0
timm==0.5.4
tokenizers==0.19.1
toolz==0.12.1
torch==2.2.2
torchaudio==2.2.2
torchmetrics==1.3.2
torchtext==0.17.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.40.0
triton==2.2.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.6
xxhash==3.4.1
yarl==1.9.4
